pyproject.toml
[project]
name = "taskzone"
version = "0.1.0"
description = ""
authors = [
    {name = "dimon0804",email = "dimongamedeveloper@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiogram (>=3.18.0,<4.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytz (>=2025.1,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
